split nifs
